+++ providers/v1/sakura/client/validate_test.go
@@ -1,0 +1,80 @@
+/*
+Copyright © The ESO Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+	https://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package client_test
+
+import (
+	"context"
+	"errors"
+	"testing"
+
+	v1 "github.com/sacloud/secretmanager-api-go/apis/v1"
+	"github.com/stretchr/testify/require"
+
+	esv1 "github.com/external-secrets/external-secrets/apis/externalsecrets/v1"
+	"github.com/external-secrets/external-secrets/providers/v1/sakura/client"
+	"github.com/external-secrets/external-secrets/providers/v1/sakura/client/fake"
+)
+
+func TestValidate(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name       string
+		mockSetup  func(*testing.T, *fake.MockSecretAPIClient)
+		wantResult esv1.ValidationResult
+		wantErr    bool
+	}{
+		{
+			name: "validation success",
+			mockSetup: func(t *testing.T, mc *fake.MockSecretAPIClient) {
+				mc.WithListFunc(func(_ context.Context) ([]v1.Secret, error) {
+					return []v1.Secret{}, nil
+				})
+			},
+			wantResult: esv1.ValidationResultReady,
+			wantErr:    false,
+		},
+		{
+			name: "validation failure",
+			mockSetup: func(t *testing.T, mc *fake.MockSecretAPIClient) {
+				mc.WithListFunc(func(_ context.Context) ([]v1.Secret, error) {
+					return nil, errors.New("API error")
+				})
+			},
+			wantResult: esv1.ValidationResultError,
+			wantErr:    true,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+
+			mockClient := fake.NewMockSecretAPIClient(t)
+			tt.mockSetup(t, mockClient)
+			client := client.NewClient(mockClient)
+
+			result, err := client.Validate()
+			if tt.wantErr {
+				require.Error(t, err)
+			} else {
+				require.NoError(t, err)
+			}
+			require.Equal(t, tt.wantResult, result)
+		})
+	}
+}
